Raft of changes to support Saxon 10

+++ src/main/java/com/xmlcalabash/util/RebasedNode.java
@@ -32,7 +32,7 @@ public class RebasedNode extends AbstractVirtualNode implements WrappingFunction
      * factory method
      *
      * @param node   The node to be wrapped
-     * @param parent The StrippedNode that wraps the parent of this node
+     * @param parent The RebasedNode that wraps the parent of this node
      */
 
     protected RebasedNode(NodeInfo node, RebasedNode parent) {
@@ -156,7 +156,7 @@ public class RebasedNode extends AbstractVirtualNode implements WrappingFunction
      */
 
     /*@Nullable*/
-    public AxisIterator iterateAxis(byte axisNumber) {
+    public AxisIterator iterateAxis(int axisNumber) {
         return new WrappingIterator(node.iterateAxis(axisNumber), this, null);
     }
